+++ core/src/main/scala/com/jpro/web/View.scala
@@ -14,5 +14,6 @@ abstract class View extends Result {
   def content: Node
   def saveScrollPosition = true
   def fullscreen = false
+  def onClose(): Unit = {}
 }